Replace diff-match-patch with js-diff and fix missing filesystem deletion

+++ rest-api/db/migrations/20250410150518_filesystem_triggers.sql
@@ -1,4 +1,10 @@
 -- migrate:up
+DROP TRIGGER _cleanup_filesystem_website ON internal.website;
+
+DROP TRIGGER _cleanup_filesystem_article ON internal.article;
+
+DROP FUNCTION internal.cleanup_filesystem;
+
 CREATE FUNCTION internal.cleanup_filesystem ()
   RETURNS TRIGGER
   AS $$
@@ -26,10 +32,12 @@ BEGIN
   WHERE
     u.id = _website_user_id;
   _preview_path := _base_path || '/previews/' || _website_id;
-  _prod_path := _base_path || '/' || _username || '/' || _website_slug;
   IF TG_TABLE_NAME = 'website' THEN
     EXECUTE FORMAT('COPY (SELECT 1) TO PROGRAM ''rm -rf %s''', _preview_path);
-    EXECUTE FORMAT('COPY (SELECT 1) TO PROGRAM ''rm -rf %s''', _prod_path);
+    IF _username IS NOT NULL THEN
+      _prod_path := _base_path || '/' || _username || '/' || _website_slug;
+      EXECUTE FORMAT('COPY (SELECT 1) TO PROGRAM ''rm -rf %s''', _prod_path);
+    END IF;
   ELSIF TG_TABLE_NAME = 'article' THEN
     SELECT
       a.slug INTO _article_slug
@@ -45,6 +53,23 @@ $$
 LANGUAGE plpgsql
 SECURITY DEFINER;
 
+CREATE FUNCTION internal.cleanup_user_directory ()
+  RETURNS TRIGGER
+  AS $$
+DECLARE
+  _username TEXT;
+  _base_path CONSTANT TEXT := '/var/www/archtika-websites';
+  _user_path TEXT;
+BEGIN
+  _username := OLD.username;
+  _user_path := _base_path || '/' || _username;
+  EXECUTE FORMAT('COPY (SELECT 1) TO PROGRAM ''rm -rf %s''', _user_path);
+  RETURN OLD;
+END;
+$$
+LANGUAGE plpgsql
+SECURITY DEFINER;
+
 CREATE TRIGGER _cleanup_filesystem_website
   BEFORE UPDATE OF title OR DELETE ON internal.website
   FOR EACH ROW
@@ -54,5 +79,10 @@ CREATE TRIGGER _cleanup_filesystem_article
   BEFORE UPDATE OF title OR DELETE ON internal.article
   FOR EACH ROW
   EXECUTE FUNCTION internal.cleanup_filesystem ();
+
+CREATE TRIGGER _cleanup_user_directory
+  BEFORE DELETE ON internal.user
+  FOR EACH ROW
+  EXECUTE FUNCTION internal.cleanup_user_directory ();
 
 -- migrate:down
